# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Conclave Engine",
    "description": "**Air-Gapped Synthetic Data Generation Engine** \u2014 v1.0\n\nTransforms production databases into privacy-safe synthetic replicas inside your security perimeter, on your hardware, with zero network calls out.\n\nAll business-logic endpoints are versioned under `/api/v1/`. Infrastructure endpoints (health, unseal, auth, license) remain at root.\n\n**Authentication**: All business endpoints require a JWT Bearer token. Obtain one via `POST /auth/token`.\n\n**Error format**: All error responses use [RFC 7807 Problem Details](https://www.rfc-editor.org/rfc/rfc7807).",
    "contact": {
      "name": "Conclave Engine Operations",
      "url": "https://github.com/example/conclave-engine"
    },
    "license": {
      "name": "AGPL-3.0-or-later",
      "url": "https://www.gnu.org/licenses/agpl-3.0.html"
    },
    "version": "1.0.0rc1"
  },
  "paths": {
...
    "/security/shred": {
      "post": {
        "tags": [
          "security",
          "security"
        ],
        "summary": "Emergency cryptographic shred",
        "description": "Destroy all encryption keys and artifacts. IRREVERSIBLE. Reachable even when vault is sealed.",
        "operationId": "shred_vault_security_shred_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Unauthorized \u2014 Bearer token missing, invalid, or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "type": {
                      "type": "string",
                      "description": "URI reference identifying the problem type.",
                      "example": "about:blank"
                    },
                    "title": {
                      "type": "string",
                      "description": "Short human-readable summary of the problem type.",
                      "example": "Unauthorized"
                    },
                    "status": {
                      "type": "integer",
                      "description": "HTTP status code.",
                      "example": 401
                    },
                    "detail": {
                      "type": "string",
                      "description": "Human-readable explanation specific to this occurrence.",
                      "example": "Bearer token missing or invalid."
                    }
                  },
                  "type": "object",
                  "required": [
                    "status",
                    "title",
                    "detail"
                  ]
                },
                "example": {
                  "type": "about:blank",
                  "title": "Unauthorized",
                  "status": 401,
                  "detail": "Bearer token missing or invalid."
                }
              }
            }
          },
          "403": {
            "description": "Forbidden \u2014 authenticated operator lacks required scope.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "type": {
                      "type": "string",
                      "description": "URI reference identifying the problem type.",
                      "example": "about:blank"
                    },
                    "title": {
                      "type": "string",
                      "description": "Short human-readable summary of the problem type.",
                      "example": "Unauthorized"
                    },
                    "status": {
                      "type": "integer",
                      "description": "HTTP status code.",
                      "example": 401
                    },
                    "detail": {
                      "type": "string",
                      "description": "Human-readable explanation specific to this occurrence.",
                      "example": "Bearer token missing or invalid."
                    }
                  },
                  "type": "object",
                  "required": [
                    "status",
                    "title",
                    "detail"
                  ]
                },
                "example": {
                  "type": "about:blank",
                  "title": "Forbidden",
                  "status": 403,
                  "detail": "This operation requires the 'settings:write' scope."
                }
              }
            }
          },
          "404": {
            "description": "Not found \u2014 resource does not exist or belongs to another operator.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "type": {
                      "type": "string",
                      "description": "URI reference identifying the problem type.",
                      "example": "about:blank"
                    },
                    "title": {
                      "type": "string",
                      "description": "Short human-readable summary of the problem type.",
                      "example": "Unauthorized"
                    },
                    "status": {
                      "type": "integer",
                      "description": "HTTP status code.",
                      "example": 401
                    },
                    "detail": {
                      "type": "string",
                      "description": "Human-readable explanation specific to this occurrence.",
                      "example": "Bearer token missing or invalid."
                    }
                  },
                  "type": "object",
                  "required": [
                    "status",
                    "title",
                    "detail"
                  ]
                },
                "example": {
                  "type": "about:blank",
                  "title": "Not Found",
                  "status": 404,
                  "detail": "Job 42 not found."
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable entity \u2014 request body failed validation.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "type": {
                      "type": "string",
                      "description": "URI reference identifying the problem type.",
                      "example": "about:blank"
                    },
                    "title": {
                      "type": "string",
                      "description": "Short human-readable summary of the problem type.",
                      "example": "Unauthorized"
                    },
                    "status": {
                      "type": "integer",
                      "description": "HTTP status code.",
                      "example": 401
                    },
                    "detail": {
                      "type": "string",
                      "description": "Human-readable explanation specific to this occurrence.",
                      "example": "Bearer token missing or invalid."
                    }
                  },
                  "type": "object",
                  "required": [
                    "status",
                    "title",
                    "detail"
                  ]
                },
                "example": {
                  "type": "about:blank",
                  "title": "Validation Error",
                  "status": 422,
                  "detail": "epsilon must be > 0."
                }
              }
            }
          },
          "423": {
            "description": "Locked \u2014 vault is sealed; unseal before making this request.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "type": {
                      "type": "string",
                      "description": "URI reference identifying the problem type.",
                      "example": "about:blank"
                    },
                    "title": {
                      "type": "string",
                      "description": "Short human-readable summary of the problem type.",
                      "example": "Unauthorized"
                    },
                    "status": {
                      "type": "integer",
                      "description": "HTTP status code.",
                      "example": 401
                    },
                    "detail": {
                      "type": "string",
                      "description": "Human-readable explanation specific to this occurrence.",
                      "example": "Bearer token missing or invalid."
                    }
                  },
                  "type": "object",
                  "required": [
                    "status",
                    "title",
                    "detail"
                  ]
                },
                "example": {
                  "type": "about:blank",
                  "title": "Vault Is Sealed",
                  "status": 423,
                  "detail": "Call POST /unseal with the vault passphrase to proceed."
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable \u2014 a required dependency (database, Redis) is down.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "type": {
                      "type": "string",
                      "description": "URI reference identifying the problem type.",
                      "example": "about:blank"
                    },
                    "title": {
                      "type": "string",
                      "description": "Short human-readable summary of the problem type.",
                      "example": "Unauthorized"
                    },
                    "status": {
                      "type": "integer",
                      "description": "HTTP status code.",
                      "example": 401
                    },
                    "detail": {
                      "type": "string",
                      "description": "Human-readable explanation specific to this occurrence.",
                      "example": "Bearer token missing or invalid."
                    }
                  },
                  "type": "object",
                  "required": [
                    "status",
                    "title",
                    "detail"
                  ]
                },
                "example": {
                  "type": "about:blank",
                  "title": "Service Unavailable",
                  "status": 503,
                  "detail": "Database connection unavailable."
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}